+++ packages/partitura_core/lib/src/smufl/smufl_codepoints.dart
@@ -12,6 +12,7 @@ const Map<String, String> smuflCodepoints = {
   'gClef': '\uE050',
   'fClef': '\uE062',
   'cClef': '\uE05C',
+  'unpitchedPercussionClef1': '\uE069',
   'noteheadWhole': '\uE0A2',
   'noteheadHalf': '\uE0A3',
   'noteheadBlack': '\uE0A4',